requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
demjson==2.2.4
Django==2.1.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
idna==2.9
Pillow==7.1.2
PyExecJS==1.5.1
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.2
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
